(feat) Initial approach to use embedded word cloud functions
